Implemented socket events & outgoing middleware
